feat: MLP 재편 — 스크리너·분석 페이지 신설 및 URL 구조 개선
Phase 1 — 미사용 라우트 삭제
- app/(chat), (laboratory), (report), (ads), (algorithm-trade-legacy) 디렉토리 삭제
- lib/features/ai/, sample/ 슬라이스 삭제
- lib/store.tsx에서 삭제된 슬라이스 참조 제거

Phase 2 — URL 변경 (redirect 스텁)
- /algorithm-trade → /screener (301 redirect)
- /search → /analyze (ticker 파라미터 보존)

Phase 3 — 랜딩 페이지 온보딩 개선
- 히어로 CTA: "종목 발굴하기" → /screener 연결
- 서비스 카드 순서 재배치: 스크리너 first
- How it works 3단계: 발굴 → 분석 → 수익계산 흐름으로 업데이트

Phase 4 — /screener 신규 페이지
- D1 daily scan 기반 종목 발굴 스크리너
- 전략 탭 9개 (NCAV, 저PBR, 저PER, S-RIM 등) + 전략별 종목 수 배지
- 데스크탑 테이블 / 모바일 카드 반응형 레이아웃
- 정렬·필터 기능, 종목 클릭 시 /analyze?from=screener 이동

Phase 5 — /analyze 신규 페이지
- /search 페이지 기반으로 /analyze 경로에 새 분석 페이지 생성
- 스크리너에서 진입 시 "← 스크리너로 돌아가기" 브레드크럼 표시
- 검색 push URL /search → /analyze 수정
- useStockSearch 훅에서 삭제된 AI 슬라이스 의존성 제거

Nav 업데이트
- 메뉴 항목: 종목 발굴(/screener) · 적정 주가(/analyze) · 계산기(/calculator)
- 드롭다운 CTA: "종목 발굴 시작하기" → /screener
