--- cf-idiotquant/app/(laboratory)/laboratory/edges.tsx
@@ -1,9 +1,0 @@
-
-const initialEdges = [
-    // { id: 'edge-1', source: 'date-2025-04-25', target: 'buy-삼성전자-2025-04-25', animated: true, style: { stroke: '#4ade80' } },
-    // { id: 'edge-2', source: 'date-2025-04-25', target: 'sell-현대차-2025-04-25', animated: true, style: { stroke: '#f87171' } }
-    { id: 'edge-1', source: 'date-2025-04-25', target: 'buy-삼성전자-2025-04-25' },
-    { id: 'edge-2', source: 'date-2025-04-25', target: 'sell-현대차-2025-04-25' }
-];
-
-export default initialEdges;
